requirements.txt
python-dotenv==1.0.1
pydantic==2.11.7
